refactor: address PR #49 review — one taxonomy, YAML REW, pages/ split (#49)
Land all 12 inline review comments with a net-simpler codebase:

- DAL: revert `dsp` to key on `DSPConfig.player_id`, drop `resolve_for_player`
  and the artifact docstring note (#1, #2).
- Models: adopt CamillaDSP casing on `Band.type`
  (Peaking/Lowshelf/Highshelf/Lowpass/Highpass) and delete `compiler._TYPE_MAP` (#4).
- Constants: centralize `DEFAULT_Q` (with Butterworth rationale) and `PASS_TYPES`
  in `models/dsp.py`; remove the three duplicate copies (#7, #9).
- REW: rewrite `domains/dsp/rew.py` to speak CamillaDSP YAML — the only format
  REW v5.20.14+ round-trips — replacing the regex parser; declare `pyyaml` (#8).
- Cleanup: strip `WS-*` plan refs, the over-hot load clamp, "legacy" framing,
  and the dietpi why-not comment (#6, #10, #12).
- Docs: strengthen the presets plain-json rationale in the DAL + AGENTS.md (#3).
- Split `streamer/pages.py` into a `pages/` package (`__init__`, `index`, `dsp`,
  `_plex`, `_clients`); document the pattern in `ui/AGENTS.md` (#11).
