+++ AGENTS.md
@@ -42,7 +42,7 @@ All models are `@dataclass` with `from_dict()` / `from_config()` / `to_dict()` /
 - `Player` — Snapcast client: `id, host, port, connected, volume, muted, group_id`
 - `Group` — Snapcast group: `id, name, client_ids, stream_id, muted, volume`
 - `Stream` — Plex-Amp stream: `id, name, uri, status, current_track`
-- `DSPConfig` — CamillaDSP pipeline: `id, player_id, pipeline (dict), enabled`
+- `DSPConfig` — parametric-EQ config: `player_id, preamp_db, bands, enabled` (keyed by `player_id`; the CamillaDSP pipeline is compiled from `preamp_db` + `bands` on Save)
 
 `Player.group_id` and `Group.stream_id` are empty strings (not `None`) when unassigned — required by the pytensils `'str'` DTYPES constraint.
 
@@ -50,8 +50,8 @@ The `dsp` models (`Band`/`DSPConfig`/`Preset`) are pydantic `BaseModel` (the `@d
 
 ## DAL
 
-- `dsp`, `presets`, and `settings` persist via plain `json` — the pipeline/bands dicts are too complex for DTYPES validation.
-- Config files: `~/.audera/{dsp,dsp/presets,settings}/{id}.json`
+- `dsp`, `presets`, and `settings` all persist via plain `json` (not duckdb). A `DSPConfig`'s and a `Preset`'s `bands` are nested lists of objects, which duckdb's `read_json_auto` — flat/columnar under the pytensils DTYPES constraint — cannot model. The duckdb-backed DALs were retired, so every surviving DAL is plain-json for the same reason.
+- Config files: `~/.audera/{dsp,dsp/presets,settings}/{id}.json` (`dsp` is keyed by `player_id`)
 
 ## Code style
