+++ tests/domains/dsp/test_presets.py
@@ -7,8 +7,8 @@ def test_loudness_preset_shape():
     assert len(bands) == 2
 
     low, high = bands
-    assert (low.type, low.freq, low.gain, low.q) == ('LowShelf', 90.0, 10.0, 0.7)
-    assert (high.type, high.freq, high.gain, high.q) == ('HighShelf', 8000.0, 6.0, 0.7)
+    assert (low.type, low.freq, low.gain, low.q) == ('Lowshelf', 90.0, 10.0, 0.7)
+    assert (high.type, high.freq, high.gain, high.q) == ('Highshelf', 8000.0, 6.0, 0.7)
     assert all(band.enabled for band in bands)
 
 
@@ -20,13 +20,13 @@ def test_loudness_preset_ids_are_unique():
 
 
 def test_loudness_preset_needs_headroom():
-    config = DSPConfig(id='x', bands=loudness_preset())
+    config = DSPConfig(player_id='x', bands=loudness_preset())
     assert response_peak_db(config) > 0.0
 
 
 def _source_bands() -> list[Band]:
     return [
-        Band(id='b1', type='LowShelf', freq=90.0, gain=10.0, q=0.7),
+        Band(id='b1', type='Lowshelf', freq=90.0, gain=10.0, q=0.7),
         Band(id='b2', type='Peaking', freq=1000.0, gain=-3.0, q=2.0, enabled=False),
     ]
